# Backend/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "TallerMecanico | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5290/"
    }
  ],
  "paths": {
...
    "/api/v1/repuestos": {
      "get": {
        "tags": [
          "Repuestos"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Repuesto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Repuestos"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RepuestoDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/repuestos/{id}": {
      "get": {
        "tags": [
          "Repuestos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Repuestos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RepuestoDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Repuestos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/vehiculos": {
      "get": {
        "tags": [
          "Vehiculos"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Vehiculo"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Vehiculos"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehiculoDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/vehiculos/{id}": {
      "get": {
        "tags": [
          "Vehiculos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Vehiculos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehiculoDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Vehiculos"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/ordenes-trabajo": {
      "get": {
        "tags": [
          "OrdenesTrabajo"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "OrdenesTrabajo"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrdenTrabajoDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/ordenes-trabajo/{id}": {
      "get": {
        "tags": [
          "OrdenesTrabajo"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "OrdenesTrabajo"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrdenTrabajoDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "OrdenesTrabajo"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/rol": {
      "get": {
        "tags": [
          "Rol"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Role"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Rol"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RolDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/rol/{id}": {
      "get": {
        "tags": [
          "Rol"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Rol"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RolDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Rol"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Cliente": {
        "type": "object",
        "properties": {
          "clienteId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "nombre": {
            "type": "string"
          },
          "apellido": {
            "type": "string"
          },
          "telefono": {
            "type": [
              "null",
              "string"
            ]
          },
          "correo": {
            "type": [
              "null",
              "string"
            ]
          },
          "direccion": {
            "type": [
              "null",
              "string"
            ]
          },
          "ordenesTrabajos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrdenesTrabajo"
            }
          },
          "vehiculos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Vehiculo"
            }
          }
        }
      },
...
      "DetalleRepuesto": {
        "type": "object",
        "properties": {
          "detalleRepuestoId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "ordenId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "repuestoId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "cantidad": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "precioUnitario": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "orden": {
            "$ref": "#/components/schemas/OrdenesTrabajo"
          },
          "repuesto": {
            "$ref": "#/components/schemas/Repuesto"
          }
        }
      },
...
      "Marca": {
        "type": "object",
        "properties": {
          "marcaId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "nombre": {
            "type": "string"
          },
          "modelos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Modelo"
            }
          }
        }
      },
...
      "Modelo": {
        "type": "object",
        "properties": {
          "modeloId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "nombre": {
            "type": "string"
          },
          "marcaId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "marca": {
            "$ref": "#/components/schemas/Marca"
          },
          "vehiculos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Vehiculo"
            }
          }
        }
      },
...
      "OrdenesTrabajo": {
        "type": "object",
        "properties": {
          "ordenId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "fechaIngreso": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "estado": {
            "type": "string"
          },
          "observaciones": {
            "type": [
              "null",
              "string"
            ]
          },
          "clienteId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "vehiculoId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "mecanicoId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "cliente": {
            "$ref": "#/components/schemas/Cliente"
          },
          "detalleRepuestos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DetalleRepuesto"
            }
          },
          "detalleServicios": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DetalleServicio"
            }
          },
          "mecanico": {
            "$ref": "#/components/schemas/Mecanico"
          },
          "pago": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Pago"
              }
            ]
          },
          "vehiculo": {
            "$ref": "#/components/schemas/Vehiculo"
          }
        }
      },
      "OrdenTrabajoDTO": {
        "required": [
          "fechaIngreso",
          "estado",
          "observaciones",
          "clienteId",
          "vehiculoId",
          "mecanicoId"
        ],
        "type": "object",
        "properties": {
          "fechaIngreso": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "estado": {
            "type": "string"
          },
          "observaciones": {
            "type": "string"
          },
          "clienteId": {
            "maximum": 2147483647,
            "minimum": 1,
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "vehiculoId": {
            "maximum": 2147483647,
            "minimum": 1,
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "mecanicoId": {
            "maximum": 2147483647,
            "minimum": 1,
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "Repuesto": {
        "type": "object",
        "properties": {
          "repuestoId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "nombre": {
            "type": "string"
          },
          "descripcion": {
            "type": [
              "null",
              "string"
            ]
          },
          "precio": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "stock": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detalleRepuestos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DetalleRepuesto"
            }
          }
        }
      },
      "RepuestoDTO": {
        "required": [
          "nombre",
          "descripcion",
          "precio",
          "stock"
        ],
        "type": "object",
        "properties": {
          "nombre": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string"
          },
          "descripcion": {
            "type": "string"
          },
          "precio": {
            "maximum": 79228162514264337593543950335,
            "minimum": 1,
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "format": "double"
          },
          "stock": {
            "maximum": 2147483647,
            "minimum": 0,
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "RolDTO": {
        "required": [
          "nombre",
          "descripcion"
        ],
        "type": "object",
        "properties": {
          "nombre": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string"
          },
          "descripcion": {
            "type": "string"
          }
        }
      },
      "Role": {
        "type": "object",
        "properties": {
          "rolId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "nombre": {
            "type": "string"
          },
          "descripcion": {
            "type": [
              "null",
              "string"
            ]
          },
          "usuarios": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Usuario"
            }
          }
        }
      },
...
      "Usuario": {
        "type": "object",
        "properties": {
          "usuarioId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "nombre": {
            "type": "string"
          },
          "correo": {
            "type": "string"
          },
          "passwordHash": {
            "type": "string"
          },
          "rolId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "rol": {
            "$ref": "#/components/schemas/Role"
          }
        }
      },
      "Vehiculo": {
        "type": "object",
        "properties": {
          "vehiculoId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "patente": {
            "type": "string"
          },
          "anio": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "color": {
            "type": [
              "null",
              "string"
            ]
          },
          "clienteId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "modeloId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "cliente": {
            "$ref": "#/components/schemas/Cliente"
          },
          "modelo": {
            "$ref": "#/components/schemas/Modelo"
          },
          "ordenesTrabajos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrdenesTrabajo"
            }
          }
        }
      },
      "VehiculoDTO": {
        "required": [
          "patente",
          "anio",
          "color",
          "clienteId",
          "modeloId"
        ],
        "type": "object",
        "properties": {
          "patente": {
            "maxLength": 20,
            "minLength": 0,
            "type": "string"
          },
          "anio": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "color": {
            "type": "string"
          },
          "clienteId": {
            "maximum": 2147483647,
            "minimum": 1,
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "modeloId": {
            "maximum": 2147483647,
            "minimum": 1,
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      }
    }
  },
...
}